refactor model search & tweak scitype(::Model), adding tests
