+++ test/KNN.jl
@@ -28,7 +28,10 @@ fitresult, cache, report = MLJ.update(knn, 0, fitresult, cache, X, y);
 
 info(knn)
 
-X, y = X_and_y(load_boston())
+N =100
+X = (x1=rand(N), x2=rand(N), x3=rand(N))
+y = 2X.x1  - X.x2 + 0.05*rand(N)
+
 knnM = machine(knn, X, y)
 @test_logs (:info, r"Training") fit!(knnM)
 predict(knnM, MLJ.selectrows(X, 1:10))
