+++ test/Transformers.jl
@@ -1,47 +1,31 @@
 module TestTransformer
 
 # using Revise
-using MLJ
+import MLJ
+using MLJ.Transformers
 using MLJBase
-using CSV
 using Test
 using Statistics
-using DataFrames
 using CategoricalArrays
 using Tables
 
 # selecting features
-Xtable, y = datanow()
-X = DataFrame(Xtable) # will become redunant in DataFrames 0.17.0
-namesX = names(X)
+N =100
+X = (Zn=rand(N),
+     Crim=rand(N),
+     x3=categorical(rand("yn",N)),
+     x4=categorical(rand("yn",N)))
+
+namesX = schema(X).names |> collect
 selector = FeatureSelector()
 info(selector)
 fitresult, cache, report = MLJBase.fit(selector, 1, X)
 @test fitresult == namesX
-transform(selector, fitresult, X[1:2,:])
+transform(selector, fitresult, selectrows(X, 1:2))
 selector = FeatureSelector([:Zn, :Crim])
 fitresult, cache, report = MLJBase.fit(selector, 1, X)
-@test transform(selector, fitresult, X[1:2,:]) ==
-    MLJBase.selectcols(MLJBase.selectrows(X, 1:2), [:Zn, :Crim])
-
-# # relabelling with integer transformer:
-# y = rand(Char, 50)
-# allrows = eachindex(y)
-# test = 3:37
-# to_int_hypers = ToIntTransformer()
-# info(to_int_hypers)
-# fitresult, cache, report = MLJBase.fit(to_int_hypers, 1, y)
-# # to_int = Trainable(to_int_hypers, y)
-# # fit!(to_int, allrows)
-# z = transform(to_int_hypers, fitresult, y[test])
-# @test y[test] == inverse_transform(to_int_hypers, fitresult, z)
-# to_int_hypers.map_unseen_to_minus_one = true
-# fitresult, cache, report = MLJBase.fit(to_int_hypers, 1, [1, 2, 3, 4, 3])
-# @test report.values == [1, 2, 3, 4]
-# #to_int = Trainable(to_int_hypers, [1,2,3,4])
-# #fitresult, cache, report = fit!(to_int, [1,2,3,4])
-# @test transform(to_int_hypers, fitresult, 5) == -1
-# @test transform(to_int_hypers, fitresult, [5,1])[1] == -1 
+@test transform(selector, fitresult, selectrows(X, 1:2)) ==
+    selectcols(selectrows(X, 1:2), [:Zn, :Crim])
 
 # `UnivariateStandardizer`:
 stand = UnivariateStandardizer()
@@ -52,10 +36,12 @@ fitresult, cache, report = MLJBase.fit(stand, 1, [0, 2, 4])
 @test round.(Int, inverse_transform(stand, fitresult, [-1, 1, 3])) == [0, 4, 8] 
 
 # `Standardizer`:
-X, y = X_and_y(load_reduced_ames())
-X = selectrows(X, 1:5)
-X = selectcols(X, 1:5)
-train, test = partition(eachindex(y), 0.9);
+N = 5
+X = (OverallQual=rand(UInt8, N),
+     GrLivArea=rand(N),
+     Neighborhood=categorical(rand("abc", N)),
+     x1stFlrSF=rand(N),
+     TotalBsmtSF=rand(N))
 
 # introduce a field of type `Char`:
 x1 = categorical(map(Char, (X.OverallQual |> collect)))
@@ -103,10 +89,10 @@ fitresult, cache, report = MLJBase.fit(t, 2, v)
 
 # `OneHotEncoder`
 
-X = DataFrame(name=identity.(categorical(["Ben", "John", "Mary", "John"], ordered=true)),
-              height=[1.85, 1.67, 1.5, 1.67],
-              favourite_number=categorical([7, 5, 10, 5]),
-              age=[23, 23, 14, 23])
+X = (name=categorical(["Ben", "John", "Mary", "John"], ordered=true),
+     height=[1.85, 1.67, 1.5, 1.67],
+     favourite_number=categorical([7, 5, 10, 5]),
+     age=[23, 23, 14, 23])
 
 t = OneHotEncoder()
 info(t)
@@ -119,11 +105,13 @@ Xt = transform(t, fitresult, X)
 @test Xt.height == X.height
 @test Xt.favourite_number__10 == float.([false, false, true, false])
 @test Xt.age == X.age
-@test Tables.schema(Xt).names == (:name__Ben, :name__John, :name__Mary,
-                                  :height, :favourite_number__5,
-                                  :favourite_number__7, :favourite_number__10, :age) 
+@test schema(Xt).names == (:name__Ben, :name__John, :name__Mary,
+                           :height, :favourite_number__5,
+                           :favourite_number__7, :favourite_number__10,
+                           :age) 
 
-# test that *entire* pool of categoricals is used in fit, including unseen levels:
+# test that *entire* pool of categoricals is used in fit, including
+# unseen levels:
 fitresult_small, cache, _ =
     @test_logs((:info, r"Spawning 3"),
                (:info, r"Spawning 3"),
@@ -144,7 +132,11 @@ Xt = transform(t, fitresult, X)
 @test :favourite_number__5 in MLJ.schema(Xt).names
 
 # test that one may not add new columns:
-X.gender = categorical(['M', 'M', 'F', 'M'])
+X = (name=categorical(["Ben", "John", "Mary", "John"], ordered=true),
+     height=[1.85, 1.67, 1.5, 1.67],
+     favourite_number=categorical([7, 5, 10, 5]),
+     age=[23, 23, 14, 23],
+     gender=categorical(['M', 'M', 'F', 'M']))
 @test_throws Exception transform(t, fitresult, X)
 
 #
