--- test/datasets.jl
@@ -1,13 +1,0 @@
-module TestDatasets
-
-# using Revise
-using MLJ
-using MLJBase
-using CSV
-
-load_ames()
-load_boston()
-load_iris()
-
-end # module
-true
